Почистил код (как смог), добавил ссылки (в прототипе это локалка), добавил цвета с кодебая в палетку в Editors
