+++ RDG CTF/Assets/Editor.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6c6e964ebd13c9e4686d468eb1cf8d34
+guid: a52efb0959840314faa647f0377a3573
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
