+++ RDG CTF/Assets/Editor/UI_colors.colors
@@ -14,44 +14,20 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   m_Presets:
   - m_Name: 
+    m_Color: {r: 0.15686275, g: 0.17254902, b: 0.20784314, a: 1}
+  - m_Name: 
+    m_Color: {r: 0.101960786, g: 0.10980392, b: 0.13333334, a: 1}
+  - m_Name: 
+    m_Color: {r: 0.62352943, g: 0.9372549, b: 0, a: 1}
+  - m_Name: 
     m_Color: {r: 1, g: 1, b: 1, a: 1}
   - m_Name: 
     m_Color: {r: 0.23529412, g: 0.23529412, b: 0.23137255, a: 1}
   - m_Name: 
     m_Color: {r: 0.31132078, g: 0.31132078, b: 0.31132078, a: 1}
   - m_Name: 
-    m_Color: {r: 0.3490566, g: 0.3490566, b: 0.3490566, a: 1}
+    m_Color: {r: 0.9529412, g: 0.57254905, b: 0, a: 1}
   - m_Name: 
-    m_Color: {r: 0, g: 0.69411767, b: 0.44313726, a: 1}
+    m_Color: {r: 0.15686275, g: 0.17254902, b: 0.20784314, a: 1}
   - m_Name: 
-    m_Color: {r: 0, g: 1, b: 0, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.6784314, g: 1, b: 0.18431373, a: 1}
-  - m_Name: 
-    m_Color: {r: 0, g: 0.69411767, b: 0.44313726, a: 1}
-  - m_Name: 
-    m_Color: {r: 0, g: 0.92941177, b: 0.5411765, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.15686275, g: 0.32156864, b: 0.21960784, a: 1}
-  - m_Name: 
-    m_Color: {r: 1, g: 0, b: 0, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.6509434, g: 0.003070495, b: 0.003070495, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.043137256, g: 0.3764706, b: 0.6901961, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.2509804, g: 0.63529414, b: 0.84705883, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.011764706, g: 0.023529412, b: 0.21568628, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.23529412, g: 0.02745098, b: 0.3254902, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.44705883, g: 0.015686275, b: 0.33333334, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.5686275, g: 0.039215688, b: 0.40392157, a: 1}
-  - m_Name: 
-    m_Color: {r: 0.13725491, g: 0.1764706, b: 0.24705882, a: 1}
-  - m_Name: 
-    m_Color: {r: 0, g: 0.35686275, b: 0.25490198, a: 1}
-  - m_Name: 
-    m_Color: {r: 0, g: 0.5058824, b: 0.4392157, a: 1}
+    m_Color: {r: 0.43137255, g: 0.5803922, b: 0.61960787, a: 1}
