+++ RDG CTF/UserSettings/Search.index
@@ -1,0 +1,13 @@
+{
+    "name": "Assets",
+    "roots": ["Assets"],
+    "includes": [],
+    "excludes": [],
+    "options": {
+        "types": true,
+        "properties": true,
+        "extended": false,
+        "dependencies": false
+    },
+    "baseScore": 999
+}
